[ONNX] Support per channel quantization
Extending the support for quantization with per channel quantization.
An extra attribute `axis` can be found for per channel quantized tensors,
most commonly in quantized weight of Convolution or Linear module.
The PR adds support to correctly parse the `axis` attribute, and map to
ONNX representation in `QuantizeLinear` and `DequantizeLinear`.

ghstack-source-id: b882a8986e8df33fe25e0f7b5c7bb398ecddd1fe
Pull Request resolved: https://github.com/pytorch/pytorch/pull/76002