[quant][core][gpu][feature] Implemented quantized cuda adaptive average pool2d op
Summary:
The current implementation of quantized cuda adaptive average pooling uses the following:
dequant -> fp32 adaptive average pooling -> quant. This is the same numerically as quantized adaptive average pooling. This is not the ideal implementation, as we desire to operate on the quantized values directly. However, we are currently blocked on this as we are waiting for cudnn's 8.5.0 release, which is anticipated to support adaptive average pooling. When that support is made available, we will use it directly.

Test plan:
```
python test/test_quantization.py TestQuantizedOps.test_adaptive_avg_pool
```

ghstack-source-id: 4bf4c7c29b38b6d56f8bb432a6e4cbd9858eee64
Pull Request resolved: https://github.com/pytorch/pytorch/pull/76081